% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepPsychDB.R
\docType{data}
\name{DepPsychDB}
\alias{DepPsychDB}
\title{Meta-Analytic Depression Psychotherapy Database Dataset}
\format{
An object of class \code{data.frame} with 411 rows and 20 columns.
}
\source{
Meta-Analytic Depression Psychotherapy Database, maintained by VU
Amsterdam Clinical Psychology and Psychotherapy.
}
\usage{
data(DepPsychDB)
}
\description{
Meta-Analytic Depression Psychotherapy Database Dataset
}
\author{
Pim Cuijpers
}
\keyword{datasets}
