% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_app.R
\name{open_app}
\alias{open_app}
\title{Open Meta-Analytic Depression Psychotherapy Database Web App.}
\usage{
open_app()
}
\description{
Opens the Meta-Analytic Depression Psychotherapy Database Web App online.
}
\examples{
open_app()
}
\author{
Mathias Harrer
}
