% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuicPsychDB.R
\docType{data}
\name{SuicPsychDB}
\alias{SuicPsychDB}
\title{Meta-Analytic Suicide Prevention Database Dataset}
\format{
An object of class \code{escalc} (inherits from \code{data.frame}) with 113 rows and 46 columns.
}
\source{
Meta-Analytic Suicide Prevention Database.
}
\usage{
data(SuicPsychDB)
}
\description{
Meta-Analytic Suicide Prevention Database Dataset
}
\author{
Wouter van Ballegooijen, Josine Rawee
}
\keyword{datasets}
